Adapt compression to the async API
Linux v6.15 is dropping the sync compression API
(commit fce8b8d5986b "crypto: remove obsolete 'comp' compression API"),
so we have to migrate to the async compression API.

This converts compression to the async API, with no big design changes.

While this works, I no longer have access to real compression HW, and
as far as I can tell, the SW implementations do all the work upfront
and finish immediately; so I can not really test the async waiting.
My hope is that because the code changes are simple, this *is* correct.